Move ConfirmedCasesByMunicipalitiesTable to components/index/cards/charts
