+++ components/index/cards/ConfirmedCasesByMunicipalitiesCard.vue
@@ -26,7 +26,7 @@
 <script>
 import dayjs from 'dayjs'
 import Data from '@/data/patient.json'
-import ConfirmedCasesByMunicipalitiesTable from '~/components/ConfirmedCasesByMunicipalitiesTable.vue'
+import ConfirmedCasesByMunicipalitiesTable from '@/components/index/cards/charts/ConfirmedCasesByMunicipalitiesTable.vue'
 
 export default {
   components: {
