+++ components/index/cards/charts/ConfirmedCasesByMunicipalitiesTable.vue
